Add support for reading test data from stdin
Summary:
For "predict" and "test" operation, an input filename of "-" now
causes the test data to be read from stdin. This allows other
processes to easily pipe data into fastTest.
It'd be nice if this could be supported for training too, but
the training code needs to be able to seek about in the data,
which really requires a proper file.
Closes https://github.com/facebookresearch/fastText/pull/88

Reviewed By: ajoulin

Differential Revision: D3925763

Pulled By: EdouardGrave

fbshipit-source-id: b1f293d3aad7b35cb0f0240fabb52e6248fd8caf