Adding ability to consider different granularities of context. Along with the character level representation, adding sentence and paragraph level representations.
In order to have a better control over the hidden layer, I changed in the code how the -dim option is used and added a -granularities option.
The -dim option now describes the size of a representation vector.
The -granularities option describes the number of granularities. By default, it is 1, which makes the system behave exactly the same way as before.
Set to 2, the sentence level representation is used. To 3, the paragraph level is added too.

NOTE:
The code doesn't actually work for 2 or 3 granularities yet.
First, the input weight matrix has the wrong size when granularities > 1. Second, the sentence and paragraph representations are not properly built.