Wrote function to output the embeddings of all the n-grams in a word
Summary:
For visualization and qualitative experiments, wrote a function that outputs
the vectors of all the n-grams inside a word. Prints each ngram followed by
the vector. Since for words inside the vocabulary the word itself is treated
as an n-gram, output its vector if available, else output a vector of zeros.

Reviewed By: ajoulin

Differential Revision: D4258186

fbshipit-source-id: 520f42d59123fe1c76332192c795f5ef12cb4939