Clarify pretrained vectors command
It's not clear from the docs whether the .bin or .vec files should be used with `-pretrainedVectors` option (see https://github.com/facebookresearch/fastText/issues/108). This change clarifies that the .vec should be used (the error message given by the software when the .bin file is used instead is also not very clear).