t_sigmoid overflow when nan / remove new, delete from Model
Summary:
When running
```
fasttext supervised -input dbpedia.train -output /tmp/trash -lr 10.1 -loss ns -thread 1
```
the program quickly segfaults. This is because the argument to sigmoid becomes nan and the index overflows.

This diff simply documents this fact. The check might be too costly for this to be worth it.

It might still be good to add this to our FAQ.

Reviewed By: wickedfoo

Differential Revision: D5739447

fbshipit-source-id: bdeca744e011a9f390ff157e4a205aeb37e38ebe