added FAQ on how to get reproducible results (#633)
Summary:
Hi everyone, and thanks for this wonderful library. I'm relatively new to it, and I found myself struggling a bit when trying to obtain reproducible results, e.g. in order to find the the best parameters.
I found the perfect answer in a 2016 issue here on your repo (https://github.com/facebookresearch/fastText/issues/116)  and I though it could be useful to add it to the FAQs.

I'm sending you two PR:
- this one, in which I added the FAQ
- a second one, in which I modified the description in src/args.cc for the "thread" param

Of course feel free to choose which one to keep (or eventually to trash both of them).

Thanks!
Leonardo
Pull Request resolved: https://github.com/facebookresearch/fastText/pull/633

Differential Revision: D9814563

Pulled By: EdouardGrave

fbshipit-source-id: 83e4b7a7163b9013aef144dedd9b4bd5945bafdf