Refactor model testing and metrics code. (#672)
Summary:
Recently, a diff from Celebio added a new feature "test-label" that calculates precision/recall/f1-score for every label. This is very useful feature, however, it makes FastText class overcomplicated.
I propose a refactoring of model testing and metrics calculation code. It introduces MetricsAccumulator class, which is responsible for collecting stats on a dataset and calculating final metrics. Moving this functionality to separate class allows to simplify model testing code in FastText class. The same FastText::test method can be used to compute both regular and per-label metrics. This design allows MetricsAccumulator to be extended to implement different types of metrics. As result, it would be much easier to add other kinds of metrics in the future.
Pull Request resolved: https://github.com/facebookresearch/fastText/pull/672

Reviewed By: EdouardGrave

Differential Revision: D12901046

Pulled By: Celebio

fbshipit-source-id: 9dcf10de950e7fb9179c4400570d2fd7b9b1879c