Separating loss and gradient computation from model
Summary:
This commit splits the computation of the loss and the subsequent gradient into different classes. Each Loss class implements its own logic and contains the underlying data needed for the computation.

There is a behavioural change :
- now, `NegativeSampling` also uses the sigmoid output instead of softmax output for the prediction. Before this commit, it used sigmoid for train, softmax for prediction.

We are passing many information to `Loss` classes. There are two things we should think next:
- `State` class
- `Model` classes instead of `Loss` classes

Reviewed By: EdouardGrave

Differential Revision: D13359871

fbshipit-source-id: 2f53eaafb800a9a2742817aa113af5f6bd7e282e