website : updating the tutorial, python readme and change of `Download` title to `Resources`
Summary: The supervised tutorial is updated with the url of version 0.2, a small hierarchical softmax explanation and multi-label classification handling. A few typos and trailing spaces were fixed. The diff also specifies some titles on the installation guide of the python bindings, and it changes the header title from "Download" to "Resources".

Reviewed By: EdouardGrave

Differential Revision: D13558140

fbshipit-source-id: 078fe847d75c21341068a3c43e47aa0dfb33b6f4