bugfix: autotune bucket size manually set
Summary: The bucket part of the matrix is needed only when we compute subwords or when we want to take into account the word-ngrams. Otherwise the bucket size is set to 0. However, when the autotune is activated, the decision to set the bucket size to 0 should be let to autotune strategy.

Reviewed By: EdouardGrave

Differential Revision: D17629756

fbshipit-source-id: f259b7aad562fdd10a813af368171dee57a4a569