fix compilation inconsistency between gcc and clang (#982)
Summary:
The default compiler must be clang b/c there is a discrepancy with gcc when mixing `const` and `constexpr` types.  The `const` declarations for `kUnknownBestScore` and `kCutoffLimit` in `src/autotune.h` where throwing compiler errors with gcc since they are redeclared as `constexpr` in `src/autotune.cc`.  The `autotune::` namespace was also removed from the rest of the statements in `src/autotune.cc`.

reference:
https://stackoverflow.com/questions/17074089/error-redeclaration-differs-in-constexpr
Pull Request resolved: https://github.com/facebookresearch/fastText/pull/982

Reviewed By: piotr-bojanowski

Differential Revision: D19297669

Pulled By: Celebio

fbshipit-source-id: dffa86dff12d6b6ff47b86caa922cc02e4c1348c